+++ admin/includes/functions.php
@@ -375,7 +375,7 @@ function populateTables($user_data, $settings_data) {
 	}
 	
 	// Settings
-	$settings = array('site_title'=>$settings_data['site_title'], 'description'=>'A new ReallySimpleCMS website!', 'site_url'=>$settings_data['site_url'], 'admin_email'=>$settings_data['admin_email'], 'default_user_role'=>$default_user_role, 'home_page'=>$post['home_page'], 'do_robots'=>$settings_data['do_robots'], 'site_logo'=>0, 'site_icon'=>0);
+	$settings = array('site_title'=>$settings_data['site_title'], 'description'=>'A new ReallySimpleCMS website!', 'site_url'=>$settings_data['site_url'], 'admin_email'=>$settings_data['admin_email'], 'default_user_role'=>$default_user_role, 'home_page'=>$post['home_page'], 'do_robots'=>$settings_data['do_robots'], 'site_logo'=>0, 'site_icon'=>0, 'theme_color'=>'#ededed');
 	
 	// Insert the settings into the database
 	foreach($settings as $name=>$value)
@@ -1144,18 +1144,6 @@ function getFileSize($bytes, $decimals = 1) {
 	
 	// Return the converted file size
 	return sprintf("%.{$decimals}f", $bytes / pow(1024, $factor)).' '.$multiples[(int)$factor].($factor > 0 ? 'B' : '');
-}
-
-/**
- * Format a date string.
- * @since 1.2.1[a]
- *
- * @param string $date
- * @param string $format (optional; default: 'Y-m-d H:i:s')
- * @return string
- */
-function formatDate($date, $format = 'Y-m-d H:i:s') {
-	return date_format(date_create($date), $format);
 }
 
 /**